pyproject.toml
[project]
name = "yieldex"
version = "0.1.0"
description = "Yield optimizer for stablecoins in EVM"
dependencies = [
    "requests",
    "web3",
    "supabase",
    "python-dotenv",
    "python-telegram-bot",
    "web3",
    
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

